# colserde: optimize arrow encoding for some types