--- a/pkg/col/colserde/arrowbatchconverter.go
+++ b/pkg/col/colserde/arrowbatchconverter.go
@@ -16,11 +16,11 @@ import (
 	"reflect"
 	"unsafe"
 
-	"github.com/apache/arrow/go/arrow"
 	"github.com/apache/arrow/go/arrow/array"
 	"github.com/apache/arrow/go/arrow/memory"
 	"github.com/cockroachdb/cockroach/pkg/col/coldata"
 	"github.com/cockroachdb/cockroach/pkg/col/typeconv"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/duration"
 	"github.com/cockroachdb/errors"
@@ -37,9 +37,6 @@ type ArrowBatchConverter struct {
 	builders struct {
 		// boolBuilder builds arrow bool columns as a bitmap from a bool slice.
 		boolBuilder *array.BooleanBuilder
-		// binaryBuilder builds arrow []byte columns as one []byte slice with
-		// accompanying offsets from a [][]byte slice.
-		binaryBuilder *array.BinaryBuilder
 	}
 
 	scratch struct {
@@ -58,7 +55,6 @@ type ArrowBatchConverter struct {
 func NewArrowBatchConverter(typs []*types.T) (*ArrowBatchConverter, error) {
 	c := &ArrowBatchConverter{typs: typs}
 	c.builders.boolBuilder = array.NewBooleanBuilder(memory.DefaultAllocator)
-	c.builders.binaryBuilder = array.NewBinaryBuilder(memory.DefaultAllocator, arrow.BinaryTypes.Binary)
 	c.scratch.arrowData = make([]*array.Data, len(typs))
 	c.scratch.buffers = make([][]*memory.Buffer, len(typs))
 	for i := range c.scratch.buffers {
@@ -106,8 +102,8 @@ func (c *ArrowBatchConverter) BatchToArrow(batch coldata.Batch) ([]*array.Data, 
 		}
 
 		var (
-			values  []byte
-			offsets []byte
+			values       []byte
+			offsetsBytes []byte
 
 			// dataHeader is the reflect.SliceHeader of the coldata.Vec's underlying
 			// data slice that we are casting to bytes.
@@ -118,9 +114,15 @@ func (c *ArrowBatchConverter) BatchToArrow(batch coldata.Batch) ([]*array.Data, 
 
 		switch typeconv.TypeFamilyToCanonicalTypeFamily(typ.Family()) {
 		case types.BytesFamily:
-			values = serializeBytesIntoArrow(n, &offsets, vec.Bytes())
+			var offsets []int32
+			values, offsets = vec.Bytes().ToArrowSerializationFormat(n)
+			unsafeCastOffsetsArray(offsets, &offsetsBytes)
+
 		case types.JsonFamily:
-			values = serializeBytesIntoArrow(n, &offsets, &vec.JSON().Bytes)
+			var offsets []int32
+			values, offsets = vec.JSON().Bytes.ToArrowSerializationFormat(n)
+			unsafeCastOffsetsArray(offsets, &offsetsBytes)
+
 		case types.IntFamily:
 			switch typ.Width() {
 			case 16:
@@ -138,10 +140,99 @@ func (c *ArrowBatchConverter) BatchToArrow(batch coldata.Batch) ([]*array.Data, 
 			default:
 				panic(fmt.Sprintf("unexpected int width: %d", typ.Width()))
 			}
+
 		case types.FloatFamily:
 			floats := vec.Float64()[:n]
 			dataHeader = (*reflect.SliceHeader)(unsafe.Pointer(&floats))
 			datumSize = sizeOfFloat64
+
+		case types.DecimalFamily:
+			offsets := make([]int32, 0, n+1)
+			decimals := vec.Decimal()[:n]
+			// We don't know exactly how big a decimal will serialize to. Let's
+			// estimate 16 bytes.
+			values = make([]byte, 0, n*16)
+			for i := range decimals {
+				offsets = append(offsets, int32(len(values)))
+				if nulls != nil && nulls.NullAt(i) {
+					continue
+				}
+				// See apd.Decimal.String(): we use the 'G' format string for
+				// serialization.
+				values = decimals[i].Append(values, 'G')
+			}
+			offsets = append(offsets, int32(len(values)))
+			unsafeCastOffsetsArray(offsets, &offsetsBytes)
+
+		case types.IntervalFamily:
+			offsets := make([]int32, 0, n+1)
+			intervals := vec.Interval()[:n]
+			intervalSize := sizeOfInt64 * 3
+			// TODO(jordan): we could right-size this values slice by counting up the
+			// number of nulls in the nulls bitmap first, and subtracting from n.
+			values = make([]byte, intervalSize*n)
+			var curNonNullInterval int
+			for i := range intervals {
+				offsets = append(offsets, int32(curNonNullInterval*intervalSize))
+				if nulls != nil && nulls.NullAt(i) {
+					continue
+				}
+				nanos, months, days, err := intervals[i].Encode()
+				if err != nil {
+					return nil, err
+				}
+				curSlice := values[intervalSize*curNonNullInterval : intervalSize*(curNonNullInterval+1)]
+				binary.LittleEndian.PutUint64(curSlice[0:sizeOfInt64], uint64(nanos))
+				binary.LittleEndian.PutUint64(curSlice[sizeOfInt64:sizeOfInt64*2], uint64(months))
+				binary.LittleEndian.PutUint64(curSlice[sizeOfInt64*2:sizeOfInt64*3], uint64(days))
+				curNonNullInterval++
+			}
+			values = values[:intervalSize*curNonNullInterval]
+			offsets = append(offsets, int32(len(values)))
+			unsafeCastOffsetsArray(offsets, &offsetsBytes)
+
+		case types.TimestampTZFamily:
+			offsets := make([]int32, 0, n+1)
+			timestamps := vec.Timestamp()[:n]
+			// See implementation of time.MarshalBinary.
+			const timestampSize = 14
+			values = make([]byte, 0, n*timestampSize)
+			for i := range timestamps {
+				offsets = append(offsets, int32(len(values)))
+				if nulls != nil && nulls.NullAt(i) {
+					continue
+				}
+				marshaled, err := timestamps[i].MarshalBinary()
+				if err != nil {
+					return nil, err
+				}
+				values = append(values, marshaled...)
+			}
+			offsets = append(offsets, int32(len(values)))
+			unsafeCastOffsetsArray(offsets, &offsetsBytes)
+
+		case typeconv.DatumVecCanonicalTypeFamily:
+			offsets := make([]int32, 0, n+1)
+			datums := vec.Datum().Slice(0 /* start */, n)
+			// Make a very very rough estimate of the number of bytes we'll have to
+			// allocate for the datums in this vector. This will likely be an
+			// undercount, but the estimate is better than nothing.
+			size, _ := tree.DatumTypeSize(typ)
+			values = make([]byte, 0, n*int(size))
+			for i := 0; i < n; i++ {
+				offsets = append(offsets, int32(len(values)))
+				if nulls != nil && nulls.NullAt(i) {
+					continue
+				}
+				var err error
+				values, err = datums.MarshalAt(values, i)
+				if err != nil {
+					return nil, err
+				}
+			}
+			offsets = append(offsets, int32(len(values)))
+			unsafeCastOffsetsArray(offsets, &offsetsBytes)
+
 		default:
 			panic(fmt.Sprintf("unsupported type for conversion to arrow data %s", typ))
 		}
@@ -163,8 +254,8 @@ func (c *ArrowBatchConverter) BatchToArrow(batch coldata.Batch) ([]*array.Data, 
 		// WARNING: The ordering of construction is critical.
 		c.scratch.buffers[vecIdx] = c.scratch.buffers[vecIdx][:0]
 		c.scratch.buffers[vecIdx] = append(c.scratch.buffers[vecIdx], memory.NewBufferBytes(arrowBitmap))
-		if offsets != nil {
-			c.scratch.buffers[vecIdx] = append(c.scratch.buffers[vecIdx], memory.NewBufferBytes(offsets))
+		if offsetsBytes != nil {
+			c.scratch.buffers[vecIdx] = append(c.scratch.buffers[vecIdx], memory.NewBufferBytes(offsetsBytes))
 		}
 		c.scratch.buffers[vecIdx] = append(c.scratch.buffers[vecIdx], memory.NewBufferBytes(values))
 
@@ -181,17 +272,15 @@ func (c *ArrowBatchConverter) BatchToArrow(batch coldata.Batch) ([]*array.Data, 
 	return c.scratch.arrowData, nil
 }
 
-// serializeBytesIntoArrow returns the bytes of a serialized flat bytes array. It
-// unsafely updates the passed-in offsets array.
-func serializeBytesIntoArrow(batchLength int, offsets *[]byte, bytes *coldata.Bytes) []byte {
-	values, int32Offsets := bytes.ToArrowSerializationFormat(batchLength)
+// unsafeCastOffsetsArray unsafe-casts the input offsetsBytes slice to point at
+// the data inside of the int32 offsets slice.
+func unsafeCastOffsetsArray(offsetsInt32 []int32, offsetsBytes *[]byte) {
 	// Cast int32Offsets to []byte.
-	int32Header := (*reflect.SliceHeader)(unsafe.Pointer(&int32Offsets))
-	offsetsHeader := (*reflect.SliceHeader)(unsafe.Pointer(offsets))
-	offsetsHeader.Data = int32Header.Data
-	offsetsHeader.Len = int32Header.Len * sizeOfInt32
-	offsetsHeader.Cap = int32Header.Cap * sizeOfInt32
-	return values
+	int32Header := (*reflect.SliceHeader)(unsafe.Pointer(&offsetsInt32))
+	bytesHeader := (*reflect.SliceHeader)(unsafe.Pointer(offsetsBytes))
+	bytesHeader.Data = int32Header.Data
+	bytesHeader.Len = int32Header.Len * sizeOfInt32
+	bytesHeader.Cap = int32Header.Cap * sizeOfInt32
 }
 
 // batchToArrowSpecialType checks whether the vector requires special handling
@@ -214,72 +303,6 @@ func (c *ArrowBatchConverter) batchToArrowSpecialType(
 		}
 		data.Buffers()[0] = memory.NewBufferBytes(arrowBitmap)
 		return data, nil
-
-	case types.DecimalFamily:
-		decimals := vec.Decimal()[:n]
-		for i := range decimals {
-			if nulls != nil && nulls.NullAt(i) {
-				c.builders.binaryBuilder.AppendNull()
-				continue
-			}
-			marshaled, err := decimals[i].MarshalText()
-			if err != nil {
-				return nil, err
-			}
-			c.builders.binaryBuilder.Append(marshaled)
-		}
-		return c.builders.binaryBuilder.NewBinaryArray().Data(), nil
-
-	case types.TimestampTZFamily:
-		timestamps := vec.Timestamp()[:n]
-		for i := range timestamps {
-			if nulls != nil && nulls.NullAt(i) {
-				c.builders.binaryBuilder.AppendNull()
-				continue
-			}
-			marshaled, err := timestamps[i].MarshalBinary()
-			if err != nil {
-				return nil, err
-			}
-			c.builders.binaryBuilder.Append(marshaled)
-		}
-		return c.builders.binaryBuilder.NewBinaryArray().Data(), nil
-
-	case types.IntervalFamily:
-		intervals := vec.Interval()[:n]
-		// Appending to the binary builder will copy the bytes, so it's safe to
-		// reuse a scratch bytes to encode the interval into.
-		scratchIntervalBytes := make([]byte, sizeOfInt64*3)
-		for i := range intervals {
-			if nulls != nil && nulls.NullAt(i) {
-				c.builders.binaryBuilder.AppendNull()
-				continue
-			}
-			nanos, months, days, err := intervals[i].Encode()
-			if err != nil {
-				return nil, err
-			}
-			binary.LittleEndian.PutUint64(scratchIntervalBytes[0:sizeOfInt64], uint64(nanos))
-			binary.LittleEndian.PutUint64(scratchIntervalBytes[sizeOfInt64:sizeOfInt64*2], uint64(months))
-			binary.LittleEndian.PutUint64(scratchIntervalBytes[sizeOfInt64*2:sizeOfInt64*3], uint64(days))
-			c.builders.binaryBuilder.Append(scratchIntervalBytes)
-		}
-		return c.builders.binaryBuilder.NewBinaryArray().Data(), nil
-
-	case typeconv.DatumVecCanonicalTypeFamily:
-		datums := vec.Datum().Slice(0 /* start */, n)
-		for i := 0; i < n; i++ {
-			if nulls != nil && nulls.NullAt(i) {
-				c.builders.binaryBuilder.AppendNull()
-				continue
-			}
-			marshaled, err := datums.MarshalAt(i)
-			if err != nil {
-				return nil, err
-			}
-			c.builders.binaryBuilder.Append(marshaled)
-		}
-		return c.builders.binaryBuilder.NewBinaryArray().Data(), nil
 	}
 
 	return nil, nil