--- a/pkg/col/colserde/arrowbatchconverter_test.go
+++ b/pkg/col/colserde/arrowbatchconverter_test.go
@@ -145,6 +145,7 @@ func BenchmarkArrowBatchConverter(b *testing.B) {
 		types.Decimal,
 		types.Int,
 		types.Timestamp,
+		types.Interval,
 	}
 	// numBytes corresponds 1:1 to typs and specifies how many bytes we are
 	// converting on one iteration of the benchmark for the corresponding type in
@@ -154,6 +155,7 @@ func BenchmarkArrowBatchConverter(b *testing.B) {
 		fixedLen * int64(coldata.BatchSize()),
 		0, // The number of bytes for decimals will be set below.
 		8 * int64(coldata.BatchSize()),
+		3 * 8 * int64(coldata.BatchSize()),
 		3 * 8 * int64(coldata.BatchSize()),
 	}
 	// Run a benchmark on every type we care about.