Previously, the arrow batch encoder used an inefficient method to encode
the following datatypes:

- Decimal
- Interval
- Timestamp
- Other datatypes without builtin support in vectorized

This commit improves the performance significantly for these types, by
avoiding extra allocations and copies for scratch buffers for each of
the elements in a column. This will provide marginally increased
performance for distributed queries that need to send a lot of data of
these datatypes across the network.

```
name                                                             old time/op    new time/op     delta
ArrowBatchConverter/decimal/nullFraction=0.00/BatchToArrow-32    231µs ± 1%      138µs ± 1%     -40.36%  (p=0.000 n=10+10)
ArrowBatchConverter/decimal/nullFraction=0.00/ArrowToBatch-32    379µs ± 1%      384µs ± 1%     +1.34%   (p=0.001 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/BatchToArrow-32  55.5µs ± 1%     44.5µs ± 1%    -19.80%  (p=0.000 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/ArrowToBatch-32  92.3µs ± 1%     92.2µs ± 1%    ~        (p=0.912 n=10+10)
ArrowBatchConverter/interval/nullFraction=0.00/BatchToArrow-32   26.5µs ± 1%     7.6µs ± 1%     -71.14%  (p=0.000 n=10+10)
ArrowBatchConverter/interval/nullFraction=0.00/ArrowToBatch-32   2.17µs ± 3%     2.11µs ± 3%    ~        (p=0.136 n=10+10)

name                                                             old speed      new speed       delta
ArrowBatchConverter/decimal/nullFraction=0.00/BatchToArrow-32    98.8MB/s ± 1%  165.9MB/s ± 1%  +67.90%   (p=0.000 n=10+10)
ArrowBatchConverter/decimal/nullFraction=0.00/ArrowToBatch-32    60.2MB/s ± 1%  59.5MB/s ± 1%   -1.19%    (p=0.001 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/BatchToArrow-32  443MB/s ± 1%   552MB/s ± 1%    +24.69%   (p=0.000 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/ArrowToBatch-32  266MB/s ± 1%   266MB/s ± 1%    ~         (p=0.912 n=10+10)
ArrowBatchConverter/interval/nullFraction=0.00/BatchToArrow-32   927MB/s ± 1%   3213MB/s ± 1%   +246.52%  (p=0.000 n=10+10)
ArrowBatchConverter/interval/nullFraction=0.00/ArrowToBatch-32   11.3GB/s ± 3%  11.6GB/s ± 3%     ~       (p=0.143 n=10+10)

name                                                             old alloc/op   new alloc/op    delta
ArrowBatchConverter/decimal/nullFraction=0.00/BatchToArrow-32    200kB ± 0%     102kB ± 0%      -48.87%  (p=0.001 n=8+9)
ArrowBatchConverter/decimal/nullFraction=0.00/ArrowToBatch-32    110kB ± 0%     110kB ± 0%      +0.50%   (p=0.000 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/BatchToArrow-32  64.5kB ± 0%    54.3kB ± 0%     -15.77%  (p=0.000 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/ArrowToBatch-32  164kB ± 0%     164kB ± 0%      ~        (all equal)
ArrowBatchConverter/interval/nullFraction=0.00/BatchToArrow-32   89.0kB ± 0%    29.7kB ± 0%     -66.62%  (p=0.000 n=10+10)
ArrowBatchConverter/interval/nullFraction=0.00/ArrowToBatch-32   0.00B          0.00B           ~        (all equal)

name                                                             old allocs/op  new allocs/op   delta
ArrowBatchConverter/decimal/nullFraction=0.00/BatchToArrow-32    5.36k ± 0%     2.06k ± 0%      -61.67%  (p=0.000 n=10+10)
ArrowBatchConverter/decimal/nullFraction=0.00/ArrowToBatch-32    4.97k ± 0%     5.00k ± 0%      +0.48%   (p=0.000 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/BatchToArrow-32  1.05k ± 0%     1.03k ± 0%      -1.72%   (p=0.000 n=10+10)
ArrowBatchConverter/timestamp/nullFraction=0.00/ArrowToBatch-32  3.07k ± 0%     3.07k ± 0%      ~        (all equal)
ArrowBatchConverter/interval/nullFraction=0.00/BatchToArrow-32   26.0 ± 0%      6.0 ± 0%        -76.92%  (p=0.000 n=10+10)
ArrowBatchConverter/interval/nullFraction=0.00/ArrowToBatch-32   0.00           0.00            ~        (all equal)
```

Release note (performance improvement): improve the performance for
distributed queries that need to send a lot of data of certain datatypes
across the network.